Review v1.0

+++ EducationApp.DataAccessLayer/Entities/AuthorInPrintingEdition.cs
@@ -9,9 +9,9 @@ namespace EducationApp.DataAccessLayer.Entities
 {
     public class AuthorInPrintingEdition
     {
-        public int AuthorId { get; set; }
+        public long AuthorId { get; set; }
         public Author Author { get; set; }
-        public int PrintingEditionId { get; set; }
+        public long PrintingEditionId { get; set; }
         public PrintingEdition PrintingEdition { get; set; }
     }
 }
