+++ EducationApp.DataAccessLayer/Entities/OrderItem.cs
@@ -7,7 +7,7 @@ namespace EducationApp.DataAccessLayer.Entities
         public decimal Amount { get; set; }
         public int Count { get; set; }
         public Enums.Enums.Currency Currency { get; set; }
-        public int PrintingEditionId { get; set; }
+        public long PrintingEditionId { get; set; }
         public PrintingEdition PrintingEdition { get; set; }
     }
 }
